Add basic support for scrolling task lists
There are still some issues, e.g. not handling correctly what happens
when you resize the window after scrolling close to the bottom and resize
the window until you can no longer scroll. In other words, scrolling is
disabled even though the content needs to be scrolled.

There is also a 1 cell gap between the scroll down button and lowest
position of the scrollable panel in the scroll bar.